1.不再使用langchain4j 2.全部使用springAi配置

+++ memory-chat-memory/src/main/java/com/github/hambuger/memory/chat/memory/memory/model/MemoryDTO.java
@@ -46,7 +46,7 @@ public class MemoryDTO extends BaseMemoryDTO {
     /**
      * 消息向量
      */
-    private List<Float> messageContentVector;
+    private List<Double> messageContentVector;
 
 
     /**
